+++ src/add-async.js
@@ -1,15 +1,16 @@
 module.exports = function(list) {
-    return function(values, callback, items) {
-        var valuesToAdd = values.splice(0, 100);
+    var addAsync = function(values, callback, items) {
+        var valuesToAdd = values.splice(0, 50);
         items = items || [];
         items = items.concat(list.add(valuesToAdd));
         if (values.length > 0) {
             setTimeout(function() {
                 addAsync(values, callback, items);
-            }, 10);
+            }, 1);
         } else {
             list.update();
             callback(items);
         }
     };
+    return addAsync;
 };